Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.7.2
pydantic>=2.0.0

numpy>=1.24.3
langchain==0.1.0
langchain-core>=0.1.7
websockets==12.0
rank-bm25==0.2.2
motor
firebase-admin
pyrate-limiter
httpx
